+++ src/Haskellorls/Context.hs
@@ -1,5 +1,6 @@
 module Haskellorls.Context
   ( context,
+    normalColorizedContext,
     colorizedContext,
     module Haskellorls.Context.Type,
   )
@@ -13,6 +14,13 @@ import qualified Haskellorls.WrappedText as WT
 
 context :: Node.NodeInfo -> T.Text
 context = Node.getNodeContext
+
+-- | A node SELinux context decorator for the @no@ parameter of the @LS_COLORS@.
+normalColorizedContext :: Color.LsColors -> Node.NodeInfo -> [WT.WrappedText]
+normalColorizedContext lscolors node = [Color.toWrappedText lscolors getter cont]
+  where
+    cont = context node
+    getter = Color.normal
 
 colorizedContext :: Color.LsColors -> Node.NodeInfo -> [WT.WrappedText]
 colorizedContext lscolors node = [Color.toWrappedText lscolors getter cont]