+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Some acceptable values to `--color=WHEN` option for compatibility against GNU ls.
 - Support an environment variable `LS_ICONS`, which configures the `--icons` option.
 - Support the `fi` parameter of `LS_COLORS`.
+- Support the `no` parameter of `LS_COLORS`.
 
 ### Changed